some updates of physics

+++ ChaosEngine/Property.cpp
@@ -41,8 +41,8 @@ namespace ChaosEngine {
         /* Engine */
         namespace Engine {
             Type::EngineStartupProperty* StartupProp = NULL;
-            float LastTime = 0;
-            float DeltaTime = 0;
+            long double LastTime = 0;
+            long double DeltaTime = 0;
         };
 
     }
